ai-service/requirements.txt
face_recognition
opencv-python-headless
psycopg2-binary
pgvector
flask
requests